created Employee
created employee interface, moved employee tasks to the employee class
