+++ bankDeposits.json
@@ -2,3 +2,4 @@
 {"amount":10,"transaction":"Ahmed","status":"deposited"}]
 {"amount": 26.5, "bankId": 1, "transaction": "Station Deposit - Amount: 26.5", "timestamp": 1732207952752, "status": "deposited"}
 {"amount": 5.0, "bankId": 1, "transaction": "Station Deposit - Amount: 5.0", "timestamp": 1732208600612, "status": "deposited"}
+{"amount": 2.25, "bankId": 1, "transaction": "Station Deposit - Amount: 2.25", "timestamp": 1733253485890, "status": "deposited"}
